refactored species again and calculateDwm - new class DwmCalculator
put the species specific abp equations back as something that one can
get through a method in a given species. While it seemed to make more
sense to put these in the calculator, this involved switch statements
which did not ensure that every new species would have an equation. The
lack of safety prompted me to put it back in the species class. So
AbpCalculator is gone.

DwmCalcluator is a new class however in which I have put
calculateDwm(Quadrat) instead of having it in the Quadrat class. I think
this makes more sense and I will be overloading the method to calculate
for a Stand as well.

+++ Test1/app/src/main/java/com/example/drew/test1/Tree.java
@@ -16,12 +16,12 @@ import java.util.LinkedList;
 
 public class Tree{
 	
-	private int dbh;
+	private double dbh;
 	private Species species;
 	private Integer storageFactor;
 	private MaterialType materialType;
 	
-	public Tree(int dbh, Species species){
+	public Tree(double dbh, Species species){
 		
 		this.dbh = dbh;
 		this.species = species;
@@ -30,7 +30,7 @@ public class Tree{
 	}
 	//SET//--------------------------------------------
 	
-	public void setDbh(int dbh){
+	public void setDbh(double dbh){
 		this.dbh = dbh;
 	}
 	
@@ -50,7 +50,7 @@ public class Tree{
 	
 	//GET//--------------------------------------------
 	
-	public int getDbh(){
+	public double getDbh(){
 		return this.dbh;
 	}
